+++ client/src/pages/AppBody/Home.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Home = () => {
+  return <h1>Maybe we need this page???</h1>;
+};
+
+export default Home;
