--- client/src/pages/Home.jsx
@@ -1,14 +1,0 @@
-import React from 'react';
-import HomePageJumbo from '../components/HomePageJumbo';
-import HomePageCards from '../components/HomePageCards';
-
-const Home = () => {
-    return (
-        <div>
-            <HomePageJumbo />
-            <HomePageCards />
-        </div>
-    );
-};
-
-export default Home;
