--- client/src/pages/Post.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Question = () => {
-  return (
-    <div>
-      <h1>Question/Answers Page</h1>
-    </div>
-  );
-};
-
-export default Question;
