--- client/src/pages/Profile.jsx
@@ -1,42 +1,0 @@
-import React from 'react';
-import {
-  MDBRow,
-  MDBCard,
-  MDBCardBody,
-  MDBIcon,
-  MDBCardImage,
-  MDBCardText,
-  MDBCardTitle
-} from 'mdbreact';
-
-const Profile = () => {
-  return (
-    <MDBRow>
-      <MDBCard personal className="my-5">
-        <MDBCardImage
-          top
-          src="https://mdbootstrap.com/img/Photos/Avatars/img%20(29).jpg"
-          alt="MDBCard image cap"
-        />
-        <MDBCardBody>
-          <MDBCardTitle>
-            <a href="#!" className="title-one">
-              Clara
-            </a>
-          </MDBCardTitle>
-          <p className="card-meta">Joined in 2013</p>
-          <MDBCardText>Clara is an photographer living in Madrid.</MDBCardText>
-          <hr />
-          <a href="#!" className="card-meta">
-            <span>
-              <MDBIcon icon="user" />
-              22 Friends
-            </span>
-          </a>
-        </MDBCardBody>
-      </MDBCard>
-    </MDBRow>
-  );
-};
-
-export default Profile;
