+++ client/src/pages/UserProfile.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const UserProfile = () => {
+  return <h1> Hello, I am the User Profile </h1>;
+};
+
+export default UserProfile;
